added the tinder cards: will work after auth for fb implmented
