+++ app/src/main/java/dateadog/dateadog/DogSwipeActivity.java
@@ -1,28 +1,173 @@
 package dateadog.dateadog;
 
+import android.content.Context;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import android.util.Log;
+import android.view.LayoutInflater;
 import android.view.View;
+import android.view.ViewGroup;
+import android.widget.BaseAdapter;
+import android.widget.FrameLayout;
+import android.widget.ImageView;
+import android.widget.TextView;
 
-public class DogSwipeActivity extends AppCompatActivity {
+import com.bumptech.glide.Glide;
+import dateadog.dateadog.tindercard.FlingCardListener;
+import dateadog.dateadog.tindercard.SwipeFlingAdapterView;
+
+import java.util.ArrayList;
+import java.util.List;
+
+
+public class DogSwipeActivity extends AppCompatActivity implements FlingCardListener.ActionDownInterface {
+
+    public static MyAppAdapter myAppAdapter;
+    public static ViewHolder viewHolder;
+    private ArrayList<Data> al;
+    private SwipeFlingAdapterView flingContainer;
+
+    public static void removeBackground() {
+
+
+        viewHolder.background.setVisibility(View.GONE);
+        myAppAdapter.notifyDataSetChanged();
+
+    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_dog_swipe);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
+        setContentView(R.layout.activity_swipedog);
 
-        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
+        flingContainer = (SwipeFlingAdapterView) findViewById(R.id.frame);
+
+        al = new ArrayList<>();
+        al.add(new Data("http://i.dailymail.co.uk/i/pix/2016/06/17/09/3562F5B200000578-3646283-image-m-30_1466150632154.jpg", ""));
+        al.add(new Data("http://images.meredith.com/content/dam/bhg/Images/2012/11/28/405944_10150676681556019_1918501130_n.jpg.rendition.largest.ss.jpg", ""));
+        al.add(new Data("http://i.ytimg.com/vi/PnxsTxV8y3g/maxresdefault.jpg", "But I must explain to you how all this mistaken idea of denouncing pleasure and praising pain was born and I will give you a complete account of the system, and expound the actual teachings of the great explorer of the truth, the master-builder of human happiness."));
+        al.add(new Data("http://switchboard.nrdc.org/blogs/dlashof/mission_impossible_4-1.jpg", "But I must explain to you how all this mistaken idea of denouncing pleasure and praising pain was born and I will give you a complete account of the system, and expound the actual teachings of the great explorer of the truth, the master-builder of human happiness."));
+        al.add(new Data("http://i.ytimg.com/vi/PnxsTxV8y3g/maxresdefault.jpg", "But I must explain to you how all this mistaken idea of denouncing pleasure and praising pain was born and I will give you a complete account of the system, and expound the actual teachings of the great explorer of the truth, the master-builder of human happiness."));
+
+        myAppAdapter = new MyAppAdapter(al, DogSwipeActivity.this);
+        flingContainer.setAdapter(myAppAdapter);
+        flingContainer.setFlingListener(new SwipeFlingAdapterView.onFlingListener() {
             @Override
-            public void onClick(View view) {
-                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
+            public void removeFirstObjectInAdapter() {
+
+            }
+
+            @Override
+            public void onLeftCardExit(Object dataObject) {
+                al.remove(0);
+                myAppAdapter.notifyDataSetChanged();
+                //Do something on the left!
+                //You also have access to the original object.
+                //If you want to use it just cast it (String) dataObject
+
+            }
+
+            @Override
+            public void onRightCardExit(Object dataObject) {
+
+                al.remove(0);
+                myAppAdapter.notifyDataSetChanged();
+            }
+
+            @Override
+            public void onAdapterAboutToEmpty(int itemsInAdapter) {
+
+            }
+
+            @Override
+            public void onScroll(float scrollProgressPercent) {
+
+                View view = flingContainer.getSelectedView();
+                view.findViewById(R.id.background).setAlpha(0);
+                view.findViewById(R.id.item_swipe_right_indicator).setAlpha(scrollProgressPercent < 0 ? -scrollProgressPercent : 0);
+                view.findViewById(R.id.item_swipe_left_indicator).setAlpha(scrollProgressPercent > 0 ? scrollProgressPercent : 0);
             }
         });
+
+
+        // Optionally add an OnItemClickListener
+        flingContainer.setOnItemClickListener(new SwipeFlingAdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClicked(int itemPosition, Object dataObject) {
+
+                View view = flingContainer.getSelectedView();
+                view.findViewById(R.id.background).setAlpha(0);
+
+                myAppAdapter.notifyDataSetChanged();
+            }
+        });
+
+    }
+
+    @Override
+    public void onActionDownPerform() {
+        Log.e("action", "bingo");
+    }
+
+    public static class ViewHolder {
+        public static FrameLayout background;
+        public TextView DataText;
+        public ImageView cardImage;
+
+
+    }
+
+    public class MyAppAdapter extends BaseAdapter {
+
+
+        public List<Data> parkingList;
+        public Context context;
+
+        private MyAppAdapter(List<Data> apps, Context context) {
+            this.parkingList = apps;
+            this.context = context;
+        }
+
+        @Override
+        public int getCount() {
+            return parkingList.size();
+        }
+
+        @Override
+        public Object getItem(int position) {
+            return position;
+        }
+
+        @Override
+        public long getItemId(int position) {
+            return position;
+        }
+
+        @Override
+        public View getView(final int position, View convertView, ViewGroup parent) {
+
+            View rowView = convertView;
+
+
+            if (rowView == null) {
+
+                LayoutInflater inflater = getLayoutInflater();
+                rowView = inflater.inflate(R.layout.item, parent, false);
+                // configure view holder
+                viewHolder = new ViewHolder();
+                viewHolder.DataText = (TextView) rowView.findViewById(R.id.bookText);
+                viewHolder.background = (FrameLayout) rowView.findViewById(R.id.background);
+                viewHolder.cardImage = (ImageView) rowView.findViewById(R.id.cardImage);
+                rowView.setTag(viewHolder);
+
+            } else {
+                viewHolder = (ViewHolder) convertView.getTag();
+            }
+            viewHolder.DataText.setText(parkingList.get(position).getDescription() + "");
+
+            Glide.with(DogSwipeActivity.this).load(parkingList.get(position).getImagePath()).into(viewHolder.cardImage);
+
+            return rowView;
+        }
     }
 }
