+++ app/src/main/res/drawable/tinder_card_background.png
