--- app/src/main/res/layout/activity_tinderswipe.xml
@@ -1,30 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".TinderSwipeActivity">
-
-
-    <!-- activity view -->
-    <FrameLayout
-        android:layout_width="match_parent"
-
-        android:layout_height="match_parent"
-        android:background="@drawable/background">
-
-
-        <com.nkdroid.tinderswipe.tindercard.SwipeFlingAdapterView
-            android:id="@+id/frame"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@drawable/background"
-            app:rotation_degrees="15.5" />
-
-    </FrameLayout>
-
-</LinearLayout>
-
-
-
